+++ packages/plugin-session-replay-browser/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.32.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.32.1...@amplitude/plugin-session-replay-browser@1.32.2) (2026-06-22)
+
+**Note:** Version bump only for package @amplitude/plugin-session-replay-browser
+
+
+
+
+
+## [1.32.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.32.0...@amplitude/plugin-session-replay-browser@1.32.1) (2026-06-13)
+
+**Note:** Version bump only for package @amplitude/plugin-session-replay-browser
+
+
+
+
+
+# [1.32.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.31.2...@amplitude/plugin-session-replay-browser@1.32.0) (2026-06-12)
+
+
+### Bug Fixes
+
+* **session-replay-browser:** add request timeout/AbortController to send path ([#1808](https://github.com/amplitude/Amplitude-TypeScript/issues/1808)) ([1f9311e](https://github.com/amplitude/Amplitude-TypeScript/commit/1f9311e09d79a9172e35fb00911b384e5a7ed158))
+* **session-replay-browser:** coalesce style attribute mutations when mergeMutations is enabled ([#1798](https://github.com/amplitude/Amplitude-TypeScript/issues/1798)) ([36e16e2](https://github.com/amplitude/Amplitude-TypeScript/commit/36e16e2acfd4822b476ea130ce86587c423dc611))
+
+
+### Features
+
+* **session-replay-browser:** add eagerFullSnapshotSend, captureFullSnapshotOnFocus, and max event size config options ([#1806](https://github.com/amplitude/Amplitude-TypeScript/issues/1806)) ([639f911](https://github.com/amplitude/Amplitude-TypeScript/commit/639f911ed840c0ff370b5e0af278bb79e02e3ac5))
+
+
+
+
+
+## [1.31.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.31.1...@amplitude/plugin-session-replay-browser@1.31.2) (2026-06-11)
+
+**Note:** Version bump only for package @amplitude/plugin-session-replay-browser
+
+
+
+
+
 ## [1.31.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.31.0...@amplitude/plugin-session-replay-browser@1.31.1) (2026-06-05)
 
 **Note:** Version bump only for package @amplitude/plugin-session-replay-browser